Seup reading from appropriate session cookies.
The cookies are read based upon the connection type so there are secure
and insecure cookies to handle user information. Other cookies aren't as
strictly related to logging in so expandability isn't that much of an
issue.

Handling the PHP session still needs to be figured out but that could
simply be left for PHP as the information it's going to provide is
minimal, however I could use serialised arrays in the database to store
the information for user sessions in all cases rather than just when a
user is logged in and create a different session veriable for use within
the system.