Initial setup of actual working login system.
Still got a lot of work to do on it. But logging in and permissions do
work properly now. Need to be able to sign out and sessions cookies also
need to be deleted if they're found to be invalid, otherwise the system
will try to get session data every time a page is requested, which is
bad!