New feed setup. JS loaded via PHP.
The new feed setup is going to allow a more flexible way of handling
feeds, easier maintainability and extendability. JS is loaded via PHP so
that it can be loaded asynchronously on the page to decrease the display
time of a page and also reduce the number of requests sent to the server
each page load.